Move FiniteDifferences support to package extension
