+++ test/ad.jl
@@ -1,7 +1,6 @@
 using ChainRulesCore
 using ChainRulesTestUtils
 using Random
-using FiniteDifferences
 using LinearAlgebra
 
 const _repartition = @static if isdefined(Base, :get_extension)
@@ -21,30 +20,6 @@ function ChainRulesTestUtils.test_approx(actual::AbstractTensorMap,
         ChainRulesTestUtils.@test_msg msg isapprox(b, block(expected, c); kwargs...)
     end
 end
-function FiniteDifferences.to_vec(t::T) where {T<:TensorKit.TrivialTensorMap}
-    vec, from_vec = to_vec(t.data)
-    return vec, x -> T(from_vec(x), codomain(t), domain(t))
-end
-function FiniteDifferences.to_vec(t::AbstractTensorMap)
-    vec = mapreduce(vcat, blocks(t); init=scalartype(t)[]) do (c, b)
-        return reshape(b, :) .* sqrt(dim(c))
-    end
-    vec_real = scalartype(t) <: Real ? vec : collect(reinterpret(real(scalartype(t)), vec))
-
-    function from_vec(x_real)
-        x = scalartype(t) <: Real ? x_real : reinterpret(scalartype(t), x_real)
-        t′ = similar(t)
-        ctr = 0
-        for (c, b) in blocks(t′)
-            n = length(b)
-            copyto!(b, reshape(view(x, ctr .+ (1:n)), size(b)) ./ sqrt(dim(c)))
-            ctr += n
-        end
-        return t′
-    end
-    return vec_real, from_vec
-end
-FiniteDifferences.to_vec(t::TensorKit.AdjointTensorMap) = to_vec(copy(t))
 
 function _randomize!(a::TensorMap)
     for b in values(blocks(a))
